Add files via upload
In the previous commit, forgot to note that getRVinsertIndices of masks2contours_util.py, np.argmax() is now called on distances (which is correct) instead of on largeDists (which was incorrect). A minor cosmetic change has been made to getRVinsertIndices in this commit, but no major error has been fixed.